Refactor setup by using a Thor::Group class.
